Приложение переработано для работы в оффлайн-режиме

+++ app/src/main/java/com/gena_korobeynikov/yandexfinance/ui/viewModels/transactions/CreateTransactionViewModel.kt
@@ -36,7 +36,7 @@ class CreateTransactionViewModel @Inject  constructor(
     }
 
     fun createTransaction(
-        categoryId: Int,
+        categoryId: Long,
         amount: String,
         date: LocalDate,
         time: LocalTime,
